将所有实体的ID(主键)属性都改为Long型；所有入口方法的适配器参数类型都改为非基本类型；并修正由此造成的错误

+++ 01Project/dolp/02DevelopmentArea/04Code/dolp/src/com/dolplay/dolpbase/common/service/JqgridService.java
@@ -132,7 +132,7 @@ public abstract class JqgridService<T> extends IdEntityService<T> {
 			public Object invoke(Connection conn, ResultSet rs, Sql sql) throws SQLException {
 				List<StandardJqgridResDataRow> rows = new ArrayList<StandardJqgridResDataRow>();
 				int columnCount = rs.getMetaData().getColumnCount();
-				int i = 1;
+				long i = 1;
 				while (rs.next()) {
 					StandardJqgridResDataRow row = new StandardJqgridResDataRow();
 					row.setId(i);
